[084] adding navbar buttons for cardgen, moving useless funcs out, adding colorthief
