+++ front-end/src/pages/Home/Home.css
@@ -1,10 +1,19 @@
 #home {
     div.navbar.home {
         width: 50%;
+        flex-direction: column;
+        gap: .5rem;
 
-        button {
-            height: 100px;
-            width: -webkit-fill-available;
+        .navbar-row {
+            flex-direction: row;
+            gap: .5rem;
+            justify-content: center;
+
+            button {
+                height: 5rem;
+                width: 100%;
+                max-width: 50%;
+            }
         }
     }
 
@@ -23,6 +32,7 @@
 
             .stat-title {
                 width: 60%;
+                margin: .5rem 0;
 
                 text-align: center;
                 font-size: larger;
